+++ src/helpers.cpp
@@ -19,8 +19,8 @@
 using namespace std;
 using namespace Eigen;
 
-MatrixXd loadFile(string FileName, int row, int col) {
-    MatrixXd Matrix(row, col);
+MatrixXf loadFile(string FileName, int row, int col) {
+    MatrixXf Matrix(row, col);
 	ifstream File(FileName);
 	for (int k = 0; k < row; ++k) {
 		string line;
@@ -34,7 +34,7 @@ MatrixXd loadFile(string FileName, int row, int col) {
     return Matrix;
 }
 
-void saveToFile(MatrixXd &data, int rows, int cols, string FileName) {
+void saveToFile(MatrixXf &data, int rows, int cols, string FileName) {
     ofstream File; File.open(FileName);
     for (int k = 0; k < rows; ++k) {
         string line = "";
@@ -48,7 +48,7 @@ void saveToFile(MatrixXd &data, int rows, int cols, string FileName) {
     File.close();
 }
 
-vector<int> NotNanIndex(MatrixXd &A, int n) {
+vector<int> NotNanIndex(MatrixXf &A, int n) {
     vector<int> indices;
     for (int i = 0; i < n; ++i) {
         if (!isnan(A(i, 0)))
@@ -57,7 +57,7 @@ vector<int> NotNanIndex(MatrixXd &A, int n) {
     return indices;
 }
 
-void convertMatrixtoQpArray(qpOASES::real_t *qpArray, MatrixXd &Matrix, int m, int n) {
+void convertMatrixtoQpArray(qpOASES::real_t *qpArray, MatrixXf &Matrix, int m, int n) {
     int k = 0;
     for (int i = 0; i < m; ++i) {
         for (int j=0; j < n; ++j) {
